Refactor constants to be in root of public.
