+++ x-pack/plugins/index_lifecycle_management/public/constants/index.js
