+++ x-pack/plugins/index_lifecycle_management/public/sections/edit_policy/components/cold_phase/cold_phase.container.js
@@ -17,7 +17,7 @@ import {
   PHASE_COLD,
   PHASE_HOT,
   PHASE_ROLLOVER_ENABLED
-} from '../../../../store/constants';
+} from '../../../../constants';
 
 export const ColdPhase = connect(
   (state) => ({
