+++ x-pack/plugins/index_lifecycle_management/public/sections/edit_policy/components/hot_phase/hot_phase.container.js
@@ -11,7 +11,7 @@ import { connect } from 'react-redux';
 import { HotPhase as PresentationComponent } from './hot_phase';
 import { getPhase } from '../../../../store/selectors';
 import { setPhaseData } from '../../../../store/actions';
-import { PHASE_HOT, PHASE_WARM, WARM_PHASE_ON_ROLLOVER } from '../../../../store/constants';
+import { PHASE_HOT, PHASE_WARM, WARM_PHASE_ON_ROLLOVER } from '../../../../constants';
 
 export const HotPhase = connect(
   state => ({
