+++ x-pack/plugins/index_lifecycle_management/public/sections/edit_policy/components/node_allocation/node_allocation.js
@@ -7,7 +7,7 @@
 import React, { Component, Fragment } from 'react';
 import { EuiSelect, EuiButtonEmpty, EuiCallOut, EuiSpacer, EuiLoadingSpinner } from '@elastic/eui';
 import { FormattedMessage, injectI18n } from '@kbn/i18n/react';
-import { PHASE_NODE_ATTRS } from '../../../../store/constants';
+import { PHASE_NODE_ATTRS } from '../../../../constants';
 import { ErrableFormRow } from '../../form_errors';
 import { LearnMoreLink } from '../../../components/learn_more_link';
 const learnMoreLinks = (
