+++ x-pack/plugins/index_lifecycle_management/public/sections/edit_policy/components/set_priority_input.js
@@ -12,7 +12,7 @@ import {
 } from '@elastic/eui';
 import {
   PHASE_INDEX_PRIORITY,
-} from '../../../store/constants';
+} from '../../../constants';
 import { ErrableFormRow } from '../form_errors';
 import { FormattedMessage } from '@kbn/i18n/react';
 export const SetPriorityInput = props => {
