+++ x-pack/plugins/index_lifecycle_management/public/sections/edit_policy/components/warm_phase/warm_phase.container.js
@@ -13,7 +13,7 @@ import {
   getPhase,
 } from '../../../../store/selectors';
 import { setPhaseData } from '../../../../store/actions';
-import { PHASE_WARM, PHASE_HOT, PHASE_ROLLOVER_ENABLED } from '../../../../store/constants';
+import { PHASE_WARM, PHASE_HOT, PHASE_ROLLOVER_ENABLED } from '../../../../constants';
 
 export const WarmPhase = connect(
   state => ({
