+++ x-pack/plugins/index_lifecycle_management/public/sections/edit_policy/components/warm_phase/warm_phase.js
@@ -26,7 +26,7 @@ import {
   PHASE_PRIMARY_SHARD_COUNT,
   PHASE_REPLICA_COUNT,
   PHASE_SHRINK_ENABLED,
-} from '../../../../store/constants';
+} from '../../../../constants';
 import { SetPriorityInput } from '../set_priority_input';
 import { NodeAllocation } from '../node_allocation';
 import { ErrableFormRow } from '../../form_errors';
