+++ x-pack/plugins/index_lifecycle_management/public/store/selectors/policies.js
@@ -32,7 +32,7 @@ import {
   PHASE_FREEZE_ENABLED,
   PHASE_INDEX_PRIORITY,
   PHASE_ROLLOVER_MAX_DOCUMENTS
-} from '../constants';
+} from '../../constants';
 import {
   defaultEmptyDeletePhase,
   defaultEmptyColdPhase,
